refactor: dictionary 모듈 repo 분리 (#17)
