+++ src/api/dictionary/dictionary.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
 
+import { DictionaryRepo } from '@/repository/dictionary.repo';
+
 import { DictionaryController } from './dictionary.controller';
 import { DictionaryService } from './dictionary.service';
 
 @Module({
   controllers: [DictionaryController],
-  providers: [DictionaryService],
+  providers: [DictionaryService, DictionaryRepo],
 })
 export class DictionaryModule {}
